Add payment and subscription system

+++ LTS/LTS.API/Domain/Entities/WalletTransaction.cs
@@ -6,14 +6,13 @@ namespace LTS.API.Domain.Entities
     {
         public Guid Id { get; set; }
         public WalletTransactionType Type { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public string Description { get; set; } = string.Empty;
         public Guid? PaymentRequestId { get; set; }
         public Guid? OrganizationId { get; set; }
         public DateTime TransactionDate { get; set; }
         public string RecordedBy { get; set; } = string.Empty;
 
-        // Navigation
         public PaymentRequest? PaymentRequest { get; set; }
     }
 }
